Move load balance test to norace
